backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
Pillow==10.3.0
django-simple-history==3.7.0
Faker==25.2.0
